COMMERCE-9437 batch-planner-rest-impl - extract common code
