+++ modules/apps/batch-planner/batch-planner-rest-impl/src/main/java/com/liferay/batch/planner/rest/internal/resource/v1_0/FieldResourceImpl.java
@@ -15,25 +15,13 @@
 package com.liferay.batch.planner.rest.internal.resource.v1_0;
 
 import com.liferay.batch.planner.rest.dto.v1_0.Field;
+import com.liferay.batch.planner.rest.internal.helper.FieldHelper;
 import com.liferay.batch.planner.rest.resource.v1_0.FieldResource;
-import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.util.GetterUtil;
-import com.liferay.portal.kernel.util.ListUtil;
-import com.liferay.portal.vulcan.batch.engine.VulcanBatchEngineTaskItemDelegate;
-import com.liferay.portal.vulcan.batch.engine.VulcanBatchEngineTaskItemDelegateRegistry;
 import com.liferay.portal.vulcan.pagination.Page;
-import com.liferay.portal.vulcan.resource.OpenAPIResource;
-import com.liferay.portal.vulcan.util.OpenAPIUtil;
-import com.liferay.portal.vulcan.yaml.YAMLUtil;
-import com.liferay.portal.vulcan.yaml.openapi.OpenAPIYAML;
 
-import java.util.ArrayList;
-import java.util.Collections;
 import java.util.Comparator;
 import java.util.List;
-import java.util.Map;
-
-import javax.ws.rs.core.Response;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -53,72 +41,23 @@ public class FieldResourceImpl extends BaseFieldResourceImpl {
 			String internalClassName, Boolean export)
 		throws Exception {
 
-		List<com.liferay.portal.vulcan.batch.engine.Field> dtoEntityFields =
-			_getDTOEntityFields(internalClassName);
+		List<com.liferay.portal.vulcan.batch.engine.Field> fields =
+			_fieldHelper.getFields(internalClassName);
 
 		if (GetterUtil.getBoolean(export)) {
-			dtoEntityFields = _filterDTOEntityFields(
-				dtoEntityFields,
+			fields = _fieldHelper.filter(
+				fields,
 				com.liferay.portal.vulcan.batch.engine.Field.AccessType.WRITE);
 		}
 		else {
-			dtoEntityFields = _filterDTOEntityFields(
-				dtoEntityFields,
+			fields = _fieldHelper.filter(
+				fields,
 				com.liferay.portal.vulcan.batch.engine.Field.AccessType.READ);
 		}
 
-		dtoEntityFields.sort(
-			Comparator.comparing(dtoEntityField -> dtoEntityField.getName()));
+		fields.sort(Comparator.comparing(field -> field.getName()));
 
-		return Page.of(transform(dtoEntityFields, this::_toField));
-	}
-
-	private List<com.liferay.portal.vulcan.batch.engine.Field>
-		_filterDTOEntityFields(
-			List<com.liferay.portal.vulcan.batch.engine.Field> dtoEntityFields,
-			com.liferay.portal.vulcan.batch.engine.Field.AccessType
-				ignoredAccessType) {
-
-		return ListUtil.filter(
-			dtoEntityFields,
-			dtoEntityField -> {
-				if (dtoEntityField.getAccessType() == ignoredAccessType) {
-					return false;
-				}
-
-				String name = dtoEntityField.getName();
-
-				if (name.equals("actions") || name.startsWith("x-")) {
-					return false;
-				}
-
-				return true;
-			});
-	}
-
-	private List<com.liferay.portal.vulcan.batch.engine.Field>
-			_getDTOEntityFields(String internalClassName)
-		throws Exception {
-
-		VulcanBatchEngineTaskItemDelegate vulcanBatchEngineTaskItemDelegate =
-			_vulcanBatchEngineTaskItemDelegateRegistry.
-				getVulcanBatchEngineTaskItemDelegate(internalClassName);
-
-		Response response = _openAPIResource.getOpenAPI(
-			Collections.singleton(
-				vulcanBatchEngineTaskItemDelegate.getResourceClass()),
-			"yaml");
-
-		OpenAPIYAML openAPIYAML = YAMLUtil.loadOpenAPIYAML(
-			(String)response.getEntity());
-
-		Map<String, com.liferay.portal.vulcan.batch.engine.Field>
-			dtoEntityFields = OpenAPIUtil.getDTOEntityFields(
-				internalClassName.substring(
-					internalClassName.lastIndexOf(StringPool.PERIOD) + 1),
-				openAPIYAML);
-
-		return new ArrayList<>(dtoEntityFields.values());
+		return Page.of(transform(fields, this::_toField));
 	}
 
 	private Field _toField(com.liferay.portal.vulcan.batch.engine.Field field) {
@@ -133,10 +72,6 @@ public class FieldResourceImpl extends BaseFieldResourceImpl {
 	}
 
 	@Reference
-	private OpenAPIResource _openAPIResource;
-
-	@Reference
-	private VulcanBatchEngineTaskItemDelegateRegistry
-		_vulcanBatchEngineTaskItemDelegateRegistry;
+	private FieldHelper _fieldHelper;
 
 }
